Create Simon Says Game in C
My final assignment in my first year programming course. We were provided with an LED simulator containing different coloured lights  connected to four push buttons that could read and record data input by the user. The assignment was to create a version of the memory game "Simon Says" with five levels of difficulty. I used a series of while loops to instead design the program so that it had a potentially limitless number of playing levels.